[PACT-795] chore: clean up of script
Jira Ticket: [PACT-795]

Jira Ticket: [PACT-795]

+++ scripts/release/release-notes-onprem.sh
@@ -1,18 +1,13 @@
 #!/bin/bash
 
 # Usage: release-notes-onprem.sh -v [version] -t [tag/sha] -r 1
-
-# 0. Validate params - 
-#       version
-# 1. create branch v1.24.0-release-notes
-# 2. checkout pactflow application
-# 3. get all issues, list them in output
-# 4. create website/docs/docs/on-premises/releases/<version>.md
-# 5. update  website/sidebars.js 
-        # sed "s/\/\/on-prem-release-placeholder/\/\/on-prem-release-placeholder\n            'docs\/on-premises\/releases\/1.23.1',/" website/sidebars.js
-# 6. create website/notices/<date>-on-premises-<version>.md
-# 7. Output next steps (github PR approval)
-
+# What will this do?
+#
+# This script helps retrieve tickets that are related to the release by:
+# 1. Find relevant tickets based on commit messages
+# 2. Format the release notes 
+# 3. Create PR with release notes
+#
 
 while [[ "$#" -gt 0 ]]
   do
@@ -27,12 +22,35 @@ while [[ "$#" -gt 0 ]]
     shift
 done
 
-
-# 2. checkout application
-echo "Prepairing, please wait this may take some time..."
-echo "Fetching application..."
+####################
+# Configurtation
+####################
 DOCS_ROOT_DIT=$(pwd)
 PACTFLOW_APPLICATION_DIR=$(pwd)/scripts/release/.pactflow-application
+ONPREM_PROD_IMAGE=quay.io/pactflow/enterprise:latest
+JIRA_PROJECT_ID=17612
+JIRA_URL=https://smartbear.atlassian.net
+JIRA_USER=${JIRA_AUTH%%@*}
+
+####################
+# Validation
+####################
+if [ -z ${RELEASE_VERSION} ]; then 
+  echo "Version has not be provided."
+  exit 128
+fi
+
+if [ -z ${JIRA_AUTH} ]; then 
+  echo "JIRA_AUTH not set, please set it with by exporting JIRA_AUTH=email:token, you can create a token here: https://id.atlassian.com/manage-profile/security/api-tokens"
+  exit 128
+fi
+
+
+####################
+# Checkout Application
+####################
+echo "Prepairing, please wait, this may take some time..."
+echo "Fetching application..."
 if [ -d $PACTFLOW_APPLICATION_DIR ]; then
     rm -rf $PACTFLOW_APPLICATION_DIR
 fi
@@ -40,35 +58,34 @@ git clone git@github.com:pactflow/pactflow-application.git $PACTFLOW_APPLICATION
 cd $PACTFLOW_APPLICATION_DIR
 echo "Done."
 
-# 3. get all the issues
-ONPREM_PROD_IMAGE=quay.io/pactflow/enterprise:latest
 
+####################
+# Find SHA to diff against
+####################
 echo "Pulling latest image..."
 docker pull $ONPREM_PROD_IMAGE >/dev/null 2>&1
 PROD_TAG=$(docker inspect $ONPREM_PROD_IMAGE | jq -r '.[0].ContainerConfig.Env[] | select(startswith("PACTFLOW_GIT_SHA="))' | cut -d "=" -f2)
-DEV_TAG="HEAD"
+
+if [ -z ${DEV_TAG} ]; then 
+  DEV_TAG="HEAD"
+fi
 echo "Done."
 
-# Jira configuration in order to create version and assign tickets that will be released to it.
-JIRA_PROJECT_ID=17612
-JIRA_URL=https://smartbear.atlassian.net
-if [ -z ${JIRA_AUTH} ]; then 
-  echo "JIRA_AUTH not set, please set it with by exporting JIRA_AUTH=email:token, you can create a token here: https://id.atlassian.com/manage-profile/security/api-tokens"
-  exit 128
-fi
-JIRA_USER=${JIRA_AUTH%%@*}
 
-
+####################
+# Create branch that will merged and create version in Jira
+####################
 if [ -n "$IS_RELEASE" ]; then  
   
-  echo
-  echo
-  echo "Creating ${RELEASE_VERSION} in Jira..."
-  echo "========================"
+  echo "Creating a branch for this release..."
+  cd $DOCS_ROOT_DIT
+  git checkout -b ${RELEASE_VERSION}-release
+  cd $PACTFLOW_APPLICATION_DIR
+  echo "Done."
 
+  echo "Creating ${RELEASE_VERSION} in Jira..."
   payload_version="{\"archived\":false,\"name\":\"${RELEASE_VERSION}\",\"projectId\":${JIRA_PROJECT_ID},\"released\":false,\"description\":\"OnPrem Release for ${RELEASE_VERSION} by $JIRA_USER\"}"
   payload_issue="{\"update\": {\"fixVersions\": [{\"add\": {\"name\": \"${RELEASE_VERSION}\"}}]}}"
-
   curl -s -o /dev/null --request POST \
      --url "$JIRA_URL/rest/api/3/version" \
      --user "$JIRA_AUTH" \
@@ -78,7 +95,10 @@ if [ -n "$IS_RELEASE" ]; then
   echo "Done."
 fi
 
-### need to separate into 2 (features, fixes) and add migration notes
+
+####################
+# Find all related tickets
+####################
 fixes=""
 features=""
 migrations=""
@@ -111,7 +131,7 @@ for i in $(git log $PROD_TAG...$DEV_TAG | grep -Eo '(PACT-)([0-9]+)' | sort | un
        elif [ "$release_type" = "Fix" ] && [ "$has_note" != "null" ]; then
          fixes+="\n- "$(echo $response | jq '.fields.customfield_11009.content[].content[].text' | tr -d '"')
        else
-         review+='\n- '$i
+         review+='\n- https://smartbear.atlassian.net/browse/'$i
        fi
    fi
 
@@ -127,7 +147,9 @@ for i in $(git log $PROD_TAG...$DEV_TAG | grep -Eo '(PACT-)([0-9]+)' | sort | un
 done
 echo "Done."
 
-# 4. create website/docs/docs/on-premises/releases/<version>.md
+####################
+# Write the release notes, output to screen
+####################
 release_note_file=${DOCS_ROOT_DIT}/website/docs/docs/on-premises/releases/${RELEASE_VERSION}.md
 
 echo
@@ -149,13 +171,16 @@ ${fixes}
 ## Migration notes
 ${migrations}
 
-## Tickets to Review
-${review}
-
 " > $release_note_file
+
+echo -e "Please review the following tickets that have no release notes attached: 
+${review}"
 
 cat $release_note_file
 
+####################
+# Update things for the site
+####################
 sed "s/\/\/on-prem-release-placeholder/\/\/on-prem-release-placeholder\n            'docs\/on-premises\/releases\/$RELEASE_VERSION',/" ${DOCS_ROOT_DIT}/website/sidebars.js > ${DOCS_ROOT_DIT}/website/sidebars_temp
 mv ${DOCS_ROOT_DIT}/website/sidebars_temp ${DOCS_ROOT_DIT}/website/sidebars.js
 
@@ -169,4 +194,14 @@ tags: [on-premises, release]
 A new PactFlow on-premises release ({$RELEASE_VERSION}) is now available ([see details](/docs/on-premises/releases/{$RELEASE_VERSION})).
 " >> ${DOCS_ROOT_DIT}/website/notices/$(date +"%Y-%m-%d")-on-premises-$RELEASE_VERSION.md
 
-# 7. Output next steps (github PR approval)
+####################
+# Copy the content into the environment_variables.md file in docs.pactflow.io.
+####################
+
+####################
+# GitHub integration
+####################
+
+####################
+# Close release ??
+####################
